back_django_tts/requirements.txt
absl-py==0.14.0
appdirs==1.4.4
argcomplete==1.12.3
argon2-cffi==21.1.0
asgiref==3.4.1
astor==0.8.1
astunparse==1.6.3
attrs==21.2.0
audioread==2.1.9
backcall==0.2.0
bleach==4.1.0
boto3==1.18.53
botocore==1.21.53
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
clang==5.0
colorama==0.4.4
cycler==0.10.0
DateTime==4.3
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
Django==3.2.7
django-cors-headers==3.9.0
django-storages==1.11.1
djangorestframework==3.12.4
entrypoints==0.3
flatbuffers==1.12
gast==0.2.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.40.0
h5py==3.1.0
idna==3.2
importlib-metadata==4.8.1
inflect==5.3.0
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jamo==0.4.1
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.3
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
librosa==0.8.1
llvmlite==0.37.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.0.3
matplotlib-inline==0.1.3
mistune==0.8.4
mysqlclient==2.0.3
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
numba==0.54.0
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
pywinpty==21.0.1
pooch==1.5.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.18.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
pywin32==301
pywinpty==1.1.4
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
requests==2.26.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7.2
s3transfer==0.5.0
scikit-learn==1.0
scipy==1.7.1
Send2Trash==1.8.0
setuptools==58.0.4
six==1.16.0
SoundFile==0.10.3.post1
sqlparse==0.4.2
tensorboard==1.15.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==1.15.2
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.2
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==2.2.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
typing-extensions==3.10.0.2
Unidecode==1.3.2
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
wheel==0.37.0
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.5.0
zope.interface==5.4.0